Working on follow redirects - More code modularity - Code Broken

+++ config.go
@@ -5,7 +5,7 @@ package grab
 */
 type Config struct {
 	Url      string
-	Routines uint
+	Routines int64
 	Filename string
 	Path     string
 }
